requirements.txt
langchain==0.1.0
openai==1.6.0
python-dotenv==1.0.0
google-search-results==2.4.2
chromadb==0.4.22
tiktoken==0.5.2
